+++ src/beta/openai/audio.ts
@@ -6,6 +6,7 @@
 import type { PredictionService, PredictorService } from "../../services"
 import type { RemotePredictionService } from "../remote"
 import { SpeechService } from "./speech"
+import { TranscriptionService } from "./transcription"
 
 export class AudioService {
 
@@ -14,11 +15,17 @@ export class AudioService {
      */
     public readonly speech: SpeechService;
 
+    /**
+     * Transcribe audio.
+     */
+    private readonly transcriptions: TranscriptionService;
+
     public constructor(
         predictors: PredictorService,
         predictions: PredictionService,
         remotePredictions: RemotePredictionService
     ) {
         this.speech = new SpeechService(predictors, predictions, remotePredictions);
+        this.transcriptions = new TranscriptionService(predictors, predictions, remotePredictions);
     }
 }