+++ Changelog.md
@@ -1,4 +1,6 @@
 ## 0.0.51
++ Added `muna.beta.predictions.remote.stream` method to stream remote predictions.
++ Added support for `aac`, `flac`, `opus`, `wav` response formats in `muna.beta.openai.audio.speech.create` function.
 + Refactored `Acceleration` constants to use a `local_` prefix.
 + Removed `ParameterDenotation` type.
 + Upgraded to Function C 0.0.39.